+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 2021/05/19 - 00 - Initialized System
+> Toolbox version 0.7.1
+* Updated systems for `NJP00` and `PRL00` and updated scripts.
+* Completed system methods for `Sys00` for the modulated OEM system.
+
 ## 2021/05/16 - 00 - PRL Model
 > Toolbox version 0.7.0
 * Updated and renamed `LCNJP` model to `NJP00`.